Stun, AbilityCooldown, VictoryPage raakile, RandomItemGenerator 1.0, CombatPagen modaus näyttämään stunnin ja cooldownin.

+++ AbilityClassLibrary/NPC/Bite.cs
@@ -15,7 +15,7 @@ namespace AbilityClassLibrary.NPC
             Description = "Bites off a chunk of his enemys flesh. Tiny bit.";
         }
 
-        public int Action(int strength, int crit, double multiplier, int increase)
+        public int Action(int strength, double crit, double multiplier, int increase)
         {
             var dmg = 3 + strength;
             return AttackLogic.CalculateAttackDamage(dmg, crit, multiplier, increase);
