+++ CombatLogicClassLibrary/AttackLogic.cs
@@ -10,7 +10,7 @@ namespace CombatLogicClassLibrary
 {
     public static class AttackLogic
     {
-        public static int CalculateAttackDamage(int baseDmg, int crit, double multiplier, int increase)
+        public static int CalculateAttackDamage(int baseDmg, double crit, double multiplier, int increase)
         {
             var dmg = (baseDmg * Convert.ToInt32(multiplier)) + increase;
             dmg = RandomProvider.GetRandom(Convert.ToInt32(dmg * .8), Convert.ToInt32(dmg * 1.2));
@@ -19,7 +19,7 @@ namespace CombatLogicClassLibrary
             else return dmg;
         }
 
-        private static bool isCrit(int crit)
+        private static bool isCrit(double crit)
         {
             return crit >= RandomProvider.GetRandom(1, 100);
         }
